feat(sink): reuse grpc channel for remote sink writer
