+++ src/connector/src/lib.rs
@@ -33,6 +33,7 @@ use std::time::Duration;
 
 use duration_str::parse_std;
 use risingwave_pb::connector_service::SinkPayloadFormat;
+use risingwave_rpc_client::ConnectorClient;
 use serde::de;
 
 pub mod aws_auth;
@@ -48,17 +49,17 @@ pub mod common;
 
 #[derive(Clone, Debug, Default)]
 pub struct ConnectorParams {
-    pub connector_rpc_endpoint: Option<String>,
+    pub connector_client: Option<ConnectorClient>,
     pub sink_payload_format: SinkPayloadFormat,
 }
 
 impl ConnectorParams {
     pub fn new(
-        connector_rpc_endpoint: Option<String>,
+        connector_client: Option<ConnectorClient>,
         sink_payload_format: SinkPayloadFormat,
     ) -> Self {
         Self {
-            connector_rpc_endpoint,
+            connector_client,
             sink_payload_format,
         }
     }
