+++ src/connector/src/sink/remote.rs
@@ -19,31 +19,28 @@ use async_trait::async_trait;
 use itertools::Itertools;
 use prost::Message;
 use risingwave_common::array::StreamChunk;
-#[cfg(test)]
-use risingwave_common::catalog::Field;
 use risingwave_common::catalog::Schema;
 use risingwave_common::types::DataType;
-use risingwave_common::util::addr::HostAddr;
 use risingwave_pb::connector_service::sink_stream_request::write_batch::json_payload::RowOp;
 use risingwave_pb::connector_service::sink_stream_request::write_batch::{
     JsonPayload, Payload, StreamChunkPayload,
 };
-use risingwave_pb::connector_service::sink_stream_request::{
-    Request as SinkRequest, StartEpoch, SyncBatch, WriteBatch,
-};
 use risingwave_pb::connector_service::table_schema::Column;
-use risingwave_pb::connector_service::{
-    SinkPayloadFormat, SinkResponse, SinkStreamRequest, TableSchema,
-};
-use risingwave_rpc_client::ConnectorClient;
+use risingwave_pb::connector_service::{SinkPayloadFormat, TableSchema};
+#[cfg(test)]
+use risingwave_pb::connector_service::{SinkResponse, SinkStreamRequest};
+use risingwave_rpc_client::{ConnectorClient, SinkWriterStreamHandle};
+#[cfg(test)]
 use tokio::sync::mpsc::{UnboundedReceiver, UnboundedSender};
-use tokio_stream::StreamExt;
-use tonic::{Status, Streaming};
+#[cfg(test)]
+use tonic::Status;
+use tracing::error;
 
 use crate::sink::catalog::{SinkId, SinkType};
+use crate::sink::utils::{record_to_json, TimestampHandlingMode};
 use crate::sink::{
-    record_to_json, DummySinkCommitCoordinator, Result, Sink, SinkError, SinkWriterParam,
-    SinkWriterV1, SinkWriterV1Adapter, TimestampHandlingMode,
+    DummySinkCommitCoordinator, Result, Sink, SinkError, SinkWriterParam, SinkWriterV1,
+    SinkWriterV1Adapter,
 };
 use crate::ConnectorParams;
 
@@ -124,7 +121,7 @@ impl Sink for RemoteSink {
         ))
     }
 
-    async fn validate(&self, connector_rpc_endpoint: Option<String>) -> Result<()> {
+    async fn validate(&self, client: Option<ConnectorClient>) -> Result<()> {
         // FIXME: support struct and array in stream sink
         let columns = self.schema.fields
             .iter()
@@ -162,16 +159,11 @@ impl Sink for RemoteSink {
             })
             .collect::<Result<Vec<_>>>()?;
 
-        let address = connector_rpc_endpoint.clone().ok_or_else(|| {
-            SinkError::Remote("connector sink endpoint not specified".parse().unwrap())
-        })?;
-        let host_addr = HostAddr::try_from(&address).map_err(SinkError::from)?;
-        let client = ConnectorClient::new(host_addr).await.map_err(|err| {
-            SinkError::Remote(format!(
-                "failed to connect to connector endpoint `{}`: {:?}",
-                &address, err
-            ))
-        })?;
+        let client = client.ok_or(SinkError::Remote(
+            "connector node endpoint not specified or unable to connect to connector node"
+                .to_string(),
+        ))?;
+
         let table_schema = TableSchema {
             columns,
             pk_indices: self.pk_indices.iter().map(|i| *i as u32).collect_vec(),
@@ -191,39 +183,14 @@ impl Sink for RemoteSink {
 }
 
 #[derive(Debug)]
-enum ResponseStreamImpl {
-    Grpc(Streaming<SinkResponse>),
-    Receiver(UnboundedReceiver<SinkResponse>),
-}
-
-impl ResponseStreamImpl {
-    pub async fn next(&mut self) -> Result<SinkResponse> {
-        match self {
-            ResponseStreamImpl::Grpc(ref mut response) => response
-                .next()
-                .await
-                .unwrap_or_else(|| Err(Status::cancelled("response stream closed unexpectedly")))
-                .map_err(|e| SinkError::Remote(e.message().to_string())),
-            ResponseStreamImpl::Receiver(ref mut receiver) => {
-                receiver.recv().await.ok_or_else(|| {
-                    SinkError::Remote("response stream closed unexpectedly".to_string())
-                })
-            }
-        }
-    }
-}
-
-#[derive(Debug)]
 pub struct RemoteSinkWriter {
     pub connector_type: String,
     properties: HashMap<String, String>,
     epoch: Option<u64>,
     batch_id: u64,
     schema: Schema,
-    _client: Option<ConnectorClient>,
-    request_sender: Option<UnboundedSender<SinkStreamRequest>>,
-    response_stream: ResponseStreamImpl,
     payload_format: SinkPayloadFormat,
+    stream_handle: SinkWriterStreamHandle,
 }
 
 impl RemoteSinkWriter {
@@ -234,19 +201,10 @@ impl RemoteSinkWriter {
         connector_params: ConnectorParams,
         sink_id: SinkId,
     ) -> Result<Self> {
-        let address = connector_params.connector_rpc_endpoint.ok_or_else(|| {
-            SinkError::Remote("connector sink endpoint not specified".parse().unwrap())
-        })?;
-        let host_addr = HostAddr::try_from(&address).map_err(SinkError::from)?;
-        let client = ConnectorClient::new(host_addr).await.map_err(|err| {
-            let msg = format!(
-                "failed to connect to connector endpoint `{}`: {:?}",
-                &address, err
-            );
-            tracing::warn!(msg);
-            SinkError::Remote(msg)
-        })?;
-
+        let client = connector_params.connector_client.ok_or(SinkError::Remote(
+            "connector node endpoint not specified or unable to connect to connector node"
+                .to_string(),
+        ))?;
         let table_schema = Some(TableSchema {
             columns: schema
                 .fields()
@@ -258,8 +216,8 @@ impl RemoteSinkWriter {
                 .collect(),
             pk_indices: pk_indices.iter().map(|i| *i as u32).collect(),
         });
-        let (request_sender, mut response) = client
-            .start_sink_stream(
+        let stream_handle = client
+            .start_sink_writer_stream(
                 config.connector_type.clone(),
                 sink_id.sink_id as u64,
                 config.properties.clone(),
@@ -267,17 +225,12 @@ impl RemoteSinkWriter {
                 connector_params.sink_payload_format,
             )
             .await
-            .map_err(SinkError::from)?;
-        response.next().await.unwrap().map_err(|e| {
-            let msg = format!(
-                "failed to start sink stream for connector `{}` with error code: {}, message: {:?}",
-                &config.connector_type,
-                e.code(),
-                e.message()
-            );
-            tracing::warn!(msg);
-            SinkError::Remote(msg)
-        })?;
+            .inspect_err(|e| {
+                error!(
+                    "failed to start sink stream for connector `{}`: {:?}",
+                    &config.connector_type, e
+                )
+            })?;
         tracing::info!(
             "{:?} sink stream started with properties: {:?}",
             &config.connector_type,
@@ -290,24 +243,17 @@ impl RemoteSinkWriter {
             epoch: None,
             batch_id: 0,
             schema,
-            _client: Some(client),
-            request_sender: Some(request_sender),
-            response_stream: ResponseStreamImpl::Grpc(response),
+            stream_handle,
             payload_format: connector_params.sink_payload_format,
         })
     }
 
-    fn on_sender_alive(&mut self) -> Result<&UnboundedSender<SinkStreamRequest>> {
-        self.request_sender
-            .as_ref()
-            .ok_or_else(|| SinkError::Remote("sink has been dropped".to_string()))
-    }
-
     #[cfg(test)]
     fn for_test(
-        response_receiver: UnboundedReceiver<SinkResponse>,
+        response_receiver: UnboundedReceiver<std::result::Result<SinkResponse, Status>>,
         request_sender: UnboundedSender<SinkStreamRequest>,
     ) -> Self {
+        use risingwave_common::catalog::Field;
         let properties = HashMap::from([("output.path".to_string(), "/tmp/rw".to_string())]);
 
         let schema = Schema::new(vec![
@@ -325,15 +271,21 @@ impl RemoteSinkWriter {
             },
         ]);
 
+        use futures::StreamExt;
+        use tokio_stream::wrappers::UnboundedReceiverStream;
+
+        let stream_handle = SinkWriterStreamHandle::new(
+            request_sender,
+            UnboundedReceiverStream::new(response_receiver).boxed(),
+        );
+
         Self {
             connector_type: "file".to_string(),
             properties,
             epoch: None,
             batch_id: 0,
             schema,
-            _client: None,
-            request_sender: Some(request_sender),
-            response_stream: ResponseStreamImpl::Receiver(response_receiver),
+            stream_handle,
             payload_format: SinkPayloadFormat::Json,
         }
     }
@@ -375,47 +327,28 @@ impl SinkWriterV1 for RemoteSinkWriter {
             SinkError::Remote("epoch has not been initialize, call `begin_epoch`".to_string())
         })?;
         let batch_id = self.batch_id;
-        self.on_sender_alive()?
-            .send(SinkStreamRequest {
-                request: Some(SinkRequest::Write(WriteBatch {
-                    epoch,
-                    batch_id,
-                    payload: Some(payload),
-                })),
-            })
-            .map_err(|e| SinkError::Remote(e.to_string()))?;
-        self.response_stream
-            .next()
-            .await
-            .map(|_| self.batch_id += 1)
+        self.stream_handle
+            .write_batch(epoch, batch_id, payload)
+            .await?;
+        self.batch_id += 1;
+        Ok(())
     }
 
     async fn begin_epoch(&mut self, epoch: u64) -> Result<()> {
-        self.on_sender_alive()?
-            .send(SinkStreamRequest {
-                request: Some(SinkRequest::StartEpoch(StartEpoch { epoch })),
-            })
-            .map_err(|e| SinkError::Remote(e.to_string()))?;
-        self.response_stream
-            .next()
-            .await
-            .map(|_| self.epoch = Some(epoch))
+        self.begin_epoch(epoch).await?;
+        self.epoch = Some(epoch);
+        Ok(())
     }
 
     async fn commit(&mut self) -> Result<()> {
         let epoch = self.epoch.ok_or_else(|| {
             SinkError::Remote("epoch has not been initialize, call `begin_epoch`".to_string())
         })?;
-        self.on_sender_alive()?
-            .send(SinkStreamRequest {
-                request: Some(SinkRequest::Sync(SyncBatch { epoch })),
-            })
-            .map_err(|e| SinkError::Remote(e.to_string()))?;
-        self.response_stream.next().await.map(|_| ())
+        self.stream_handle.commit(epoch).await?;
+        Ok(())
     }
 
     async fn abort(&mut self) -> Result<()> {
-        self.request_sender = None;
         Ok(())
     }
 }
@@ -426,11 +359,9 @@ mod test {
 
     use risingwave_common::array::StreamChunk;
     use risingwave_common::test_prelude::StreamChunkTestExt;
-    use risingwave_pb::connector_service::sink_response::{
-        Response, StartEpochResponse, SyncResponse, WriteResponse,
-    };
+    use risingwave_pb::connector_service::sink_response::{Response, SyncResponse};
     use risingwave_pb::connector_service::sink_stream_request::write_batch::Payload;
-    use risingwave_pb::connector_service::sink_stream_request::Request;
+    use risingwave_pb::connector_service::sink_stream_request::{Request, SyncBatch};
     use risingwave_pb::connector_service::{SinkResponse, SinkStreamRequest};
     use risingwave_pb::data;
     use tokio::sync::mpsc;
@@ -498,11 +429,6 @@ mod test {
         );
 
         // test write batch
-        response_sender
-            .send(SinkResponse {
-                response: Some(Response::StartEpoch(StartEpochResponse { epoch: 2022 })),
-            })
-            .expect("test failed: failed to start epoch");
         sink.begin_epoch(2022).await.unwrap();
         assert_eq!(sink.epoch, Some(2022));
 
@@ -511,14 +437,6 @@ mod test {
             .await
             .expect("test failed: failed to construct start_epoch request");
 
-        response_sender
-            .send(SinkResponse {
-                response: Some(Response::Write(WriteResponse {
-                    epoch: 2022,
-                    batch_id: 0,
-                })),
-            })
-            .expect("test failed: failed to start epoch");
         sink.write_batch(chunk_a.clone()).await.unwrap();
         assert_eq!(sink.epoch, Some(2022));
         assert_eq!(sink.batch_id, 1);
@@ -548,39 +466,26 @@ mod test {
 
         // test commit
         response_sender
-            .send(SinkResponse {
+            .send(Ok(SinkResponse {
                 response: Some(Response::Sync(SyncResponse { epoch: 2022 })),
-            })
+            }))
             .expect("test failed: failed to sync epoch");
         sink.commit().await.unwrap();
         let commit_request = request_receiver.recv().await.unwrap();
         match commit_request.request {
-            Some(Request::Sync(sync_batch)) => {
-                assert_eq!(sync_batch.epoch, 2022);
+            Some(Request::Sync(SyncBatch { epoch })) => {
+                assert_eq!(epoch, 2022);
             }
             _ => panic!("test failed: failed to construct sync request "),
         }
 
         // begin another epoch
-        response_sender
-            .send(SinkResponse {
-                response: Some(Response::StartEpoch(StartEpochResponse { epoch: 2023 })),
-            })
-            .expect("test failed: failed to start epoch");
         sink.begin_epoch(2023).await.unwrap();
         // simply keep the channel empty since we've tested begin_epoch
         let _ = request_receiver.recv().await.unwrap();
         assert_eq!(sink.epoch, Some(2023));
 
         // test another write
-        response_sender
-            .send(SinkResponse {
-                response: Some(Response::Write(WriteResponse {
-                    epoch: 2022,
-                    batch_id: 1,
-                })),
-            })
-            .expect("test failed: failed to start epoch");
         sink.write_batch(chunk_b.clone()).await.unwrap();
         assert_eq!(sink.epoch, Some(2023));
         assert_eq!(sink.batch_id, 2);
