+++ src/connector/src/source/cdc/source/reader.rs
@@ -99,8 +99,7 @@ impl CdcSplitReader {
     #[try_stream(boxed, ok = Vec<SourceMessage>, error = anyhow::Error)]
     async fn into_data_stream(self) {
         tracing::debug!("cdc props: {:?}", self.conn_props);
-        let cdc_client =
-            ConnectorClient::new(HostAddr::from_str(&self.conn_props.connector_node_addr)?).await?;
+        let cdc_client = ConnectorClient::new(&self.conn_props.connector_node_addr).await?;
 
         // rewrite the hostname and port for the split
         let mut properties = self.conn_props.props.clone();
