+++ src/rpc_client/src/connector_client.rs
@@ -12,35 +12,151 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::any::type_name;
 use std::collections::HashMap;
+use std::fmt::{Debug, Formatter};
 use std::time::Duration;
 
 use anyhow::anyhow;
-use async_trait::async_trait;
+use futures::stream::BoxStream;
+use futures::StreamExt;
 use risingwave_common::config::{MAX_CONNECTION_WINDOW_SIZE, STREAM_WINDOW_SIZE};
-use risingwave_common::util::addr::HostAddr;
 use risingwave_pb::catalog::SinkType;
 use risingwave_pb::connector_service::connector_service_client::ConnectorServiceClient;
-use risingwave_pb::connector_service::sink_stream_request::{Request as SinkRequest, StartSink};
+use risingwave_pb::connector_service::sink_stream_request::write_batch::Payload;
+use risingwave_pb::connector_service::sink_stream_request::{
+    Request as SinkRequest, StartEpoch, StartSink, SyncBatch, WriteBatch,
+};
 use risingwave_pb::connector_service::*;
 use tokio::sync::mpsc::{unbounded_channel, UnboundedSender};
 use tokio_stream::wrappers::UnboundedReceiverStream;
 use tonic::transport::{Channel, Endpoint};
-use tonic::{Request, Streaming};
+use tonic::{Request, Status, Streaming};
+use tracing::error;
 
 use crate::error::{Result, RpcError};
-use crate::RpcClient;
 
 #[derive(Clone, Debug)]
-pub struct ConnectorClient(ConnectorServiceClient<Channel>);
+pub struct ConnectorClient {
+    rpc_client: ConnectorServiceClient<Channel>,
+    endpoint: String,
+}
+
+pub struct SinkWriterStreamHandle {
+    request_sender: UnboundedSender<SinkStreamRequest>,
+    response_stream: BoxStream<'static, std::result::Result<SinkResponse, Status>>,
+}
+
+impl Debug for SinkWriterStreamHandle {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        f.write_str(type_name::<Self>())
+    }
+}
+
+impl SinkWriterStreamHandle {
+    pub fn new(
+        request_sender: UnboundedSender<SinkStreamRequest>,
+        response_stream: BoxStream<'static, std::result::Result<SinkResponse, Status>>,
+    ) -> Self {
+        Self {
+            request_sender,
+            response_stream,
+        }
+    }
+
+    async fn next_response(&mut self) -> Result<SinkResponse> {
+        Ok(self
+            .response_stream
+            .next()
+            .await
+            .ok_or(RpcError::Internal(anyhow!("end of response stream")))??)
+    }
+
+    pub async fn start_epoch(&mut self, epoch: u64) -> Result<()> {
+        self.request_sender
+            .send(SinkStreamRequest {
+                request: Some(SinkRequest::StartEpoch(StartEpoch { epoch })),
+            })
+            .map_err(|_| RpcError::Internal(anyhow!("unable to send start epoch on {}", epoch)))?;
+        match self.next_response().await? {
+            SinkResponse {
+                response: Some(sink_response::Response::StartEpoch(_)),
+            } => Ok(()),
+            msg => Err(RpcError::Internal(anyhow!(
+                "should get StartEpoch response but get {:?}",
+                msg
+            ))),
+        }
+    }
+
+    pub async fn write_batch(&mut self, epoch: u64, batch_id: u64, payload: Payload) -> Result<()> {
+        self.request_sender
+            .send(SinkStreamRequest {
+                request: Some(SinkRequest::Write(WriteBatch {
+                    epoch,
+                    batch_id,
+                    payload: Some(payload),
+                })),
+            })
+            .map_err(|_| {
+                RpcError::Internal(anyhow!(
+                    "unable to send write batch on {} with batch id {}",
+                    epoch,
+                    batch_id
+                ))
+            })?;
+        match self.next_response().await? {
+            SinkResponse {
+                response: Some(sink_response::Response::Write(_)),
+            } => Ok(()),
+            msg => Err(RpcError::Internal(anyhow!(
+                "should get Write response but get {:?}",
+                msg
+            ))),
+        }
+    }
+
+    pub async fn commit(&mut self, epoch: u64) -> Result<()> {
+        self.request_sender
+            .send(SinkStreamRequest {
+                request: Some(SinkRequest::Sync(SyncBatch { epoch })),
+            })
+            .map_err(|_| RpcError::Internal(anyhow!("unable to send barrier on {}", epoch,)))?;
+        match self.next_response().await? {
+            SinkResponse {
+                response: Some(sink_response::Response::Sync(_)),
+            } => Ok(()),
+            msg => Err(RpcError::Internal(anyhow!(
+                "should get Sync response but get {:?}",
+                msg
+            ))),
+        }
+    }
+}
 
 impl ConnectorClient {
-    pub async fn new(host_addr: HostAddr) -> Result<Self> {
-        let channel = Endpoint::from_shared(format!("http://{}", &host_addr))
+    pub async fn try_connect(connector_endpoint: Option<&String>) -> Option<Self> {
+        match connector_endpoint {
+            None => None,
+            Some(connector_endpoint) => match ConnectorClient::new(connector_endpoint).await {
+                Ok(client) => Some(client),
+                Err(e) => {
+                    error!(
+                        "unable to connect to connector endpoint {:?}: {:?}",
+                        connector_endpoint, e
+                    );
+                    None
+                }
+            },
+        }
+    }
+
+    pub async fn new(connector_endpoint: &String) -> Result<Self> {
+        let channel = Endpoint::from_shared(format!("http://{}", connector_endpoint))
             .map_err(|e| {
                 RpcError::Internal(anyhow!(format!(
                     "invalid connector endpoint `{}`: {:?}",
-                    &host_addr, e
+                    &connector_endpoint, e
                 )))
             })?
             .initial_connection_window_size(MAX_CONNECTION_WINDOW_SIZE)
@@ -49,7 +165,14 @@ impl ConnectorClient {
             .connect_timeout(Duration::from_secs(5))
             .connect()
             .await?;
-        Ok(Self(ConnectorServiceClient::new(channel)))
+        Ok(Self {
+            rpc_client: ConnectorServiceClient::new(channel),
+            endpoint: connector_endpoint.to_string(),
+        })
+    }
+
+    pub fn endpoint(&self) -> &String {
+        &self.endpoint
     }
 
     /// Get source event stream
@@ -67,8 +190,8 @@ impl ConnectorClient {
             snapshot_done
         );
         Ok(self
-            .0
-            .to_owned()
+            .rpc_client
+            .clone()
             .get_event_stream(GetEventStreamRequest {
                 source_id,
                 source_type: source_type as _,
@@ -96,8 +219,8 @@ impl ConnectorClient {
         table_schema: Option<TableSchema>,
     ) -> Result<()> {
         let response = self
-            .0
-            .to_owned()
+            .rpc_client
+            .clone()
             .validate_source(ValidateSourceRequest {
                 source_id,
                 source_type: source_type as _,
@@ -118,15 +241,15 @@ impl ConnectorClient {
         })
     }
 
-    pub async fn start_sink_stream(
+    pub async fn start_sink_writer_stream(
         &self,
         connector_type: String,
         sink_id: u64,
         properties: HashMap<String, String>,
         table_schema: Option<TableSchema>,
         sink_payload_format: SinkPayloadFormat,
-    ) -> Result<(UnboundedSender<SinkStreamRequest>, Streaming<SinkResponse>)> {
-        let (request_sender, request_receiver) = unbounded_channel::<SinkStreamRequest>();
+    ) -> Result<SinkWriterStreamHandle> {
+        let (request_sender, request_receiver) = unbounded_channel();
 
         // Send initial request in case of the blocking receive call from creating streaming request
         request_sender
@@ -143,15 +266,28 @@ impl ConnectorClient {
             })
             .map_err(|err| RpcError::Internal(anyhow!(err.to_string())))?;
 
-        let response = self
-            .0
-            .to_owned()
+        let mut response = self
+            .rpc_client
+            .clone()
             .sink_stream(Request::new(UnboundedReceiverStream::new(request_receiver)))
             .await
             .map_err(RpcError::GrpcStatus)?
             .into_inner();
 
-        Ok((request_sender, response))
+        match response.next().await.ok_or(RpcError::Internal(anyhow!(
+            "get empty response from start sink request"
+        )))?? {
+            SinkResponse {
+                response: Some(sink_response::Response::Start(_)),
+            } => Ok(SinkWriterStreamHandle {
+                response_stream: response.boxed(),
+                request_sender,
+            }),
+            msg => Err(RpcError::Internal(anyhow!(
+                "should get start response but get {:?}",
+                msg
+            ))),
+        }
     }
 
     pub async fn validate_sink_properties(
@@ -162,8 +298,8 @@ impl ConnectorClient {
         sink_type: SinkType,
     ) -> Result<()> {
         let response = self
-            .0
-            .to_owned()
+            .rpc_client
+            .clone()
             .validate_sink(ValidateSinkRequest {
                 sink_config: Some(SinkConfig {
                     connector_type,
@@ -186,12 +322,5 @@ impl ConnectorClient {
                 ))))
             },
         )
-    }
-}
-
-#[async_trait]
-impl RpcClient for ConnectorClient {
-    async fn new_client(host_addr: HostAddr) -> Result<Self> {
-        Self::new(host_addr).await
     }
 }
