+++ src/source/src/source_desc.rs
@@ -129,7 +129,11 @@ impl SourceDescBuilder {
         let source = ConnectorSource::new(
             self.properties,
             columns.clone(),
-            self.connector_params.connector_rpc_endpoint,
+            // TODO: may reuse the connector client
+            self.connector_params
+                .connector_client
+                .as_ref()
+                .map(|client| client.endpoint().clone()),
             self.connector_message_buffer_size,
             psrser_config,
         )?;
@@ -161,7 +165,11 @@ impl SourceDescBuilder {
         let source = FsConnectorSource::new(
             self.properties.clone(),
             columns.clone(),
-            self.connector_params.connector_rpc_endpoint.clone(),
+            // TODO: may reuse connector client
+            self.connector_params
+                .connector_client
+                .as_ref()
+                .map(|client| client.endpoint().clone()),
             parser_config,
         )?;
